new: settings.service.ts
- new SettingsService
- deprecation warning for AlfrescoSettingsService
- using new SettingsService internally
