+++ ng2-components/ng2-alfresco-core/index.ts
@@ -24,6 +24,7 @@ import { TranslateModule, TranslateLoader } from 'ng2-translate/ng2-translate';
 import {
     AuthService,
     ContentService,
+    SettingsService,
     StorageService,
     AlfrescoApiService,
     AlfrescoSettingsService,
@@ -47,6 +48,7 @@ export * from './src/utils/index';
 export const ALFRESCO_CORE_PROVIDERS: any[] = [
     AuthService,
     ContentService,
+    SettingsService,
     StorageService,
     AlfrescoApiService,
     AlfrescoAuthenticationService,
